# Improve efficiency of the dependency graph (56% speed-up)