--- a/src/DateTimeDefault.ts
+++ b/src/DateTimeDefault.ts
@@ -6,44 +6,50 @@
 import {DateTime, SimpleDate, SimpleTime} from './DateTimeHelper'
 import {Maybe} from './Maybe'
 
+const WHITESPACE_REGEXP = new RegExp('\\s+')
+const DATE_SEPARATOR_REGEXP = new RegExp('[ /.-]')
+const TIME_SEPARATOR = ':'
+
 export function defaultParseToDateTime(dateTimeString: string, dateFormat?: string, timeFormat?: string): Maybe<DateTime> {
-  dateTimeString = dateTimeString.replace(/\s\s+/g, ' ').trim().toLowerCase()
-  let ampmtoken: Maybe<string> = dateTimeString.substring(dateTimeString.length - 2)
-  if (ampmtoken === 'am' || ampmtoken === 'pm') {
-    dateTimeString = dateTimeString.substring(0, dateTimeString.length - 2).trim()
-  } else {
-    ampmtoken = dateTimeString.substring(dateTimeString.length - 1)
-    if (ampmtoken === 'a' || ampmtoken === 'p') {
-      dateTimeString = dateTimeString.substring(0, dateTimeString.length - 1).trim()
-    } else {
-      ampmtoken = undefined
-    }
+  dateTimeString = dateTimeString.replace(WHITESPACE_REGEXP, ' ').trim().toLowerCase()
+
+  const ampmToken: Maybe<string> = extractAmpmToken(dateTimeString)
+  if (ampmToken !== undefined) {
+    dateTimeString = dateTimeString.substring(0, dateTimeString.length - ampmToken.length).trim()
+    dateTimeString = dateTimeString + TIME_SEPARATOR + ampmToken
   }
-  const dateItems = dateTimeString.split(/[ /.-]/g)
-  if (dateItems.length >= 2 && dateItems[dateItems.length - 2].includes(':')) {
+
+  const dateItems = dateTimeString.split(DATE_SEPARATOR_REGEXP)
+  if (dateItems.length >= 2 && doesLookLikeATimeString(dateItems[dateItems.length - 2])) {
     dateItems[dateItems.length - 2] = dateItems[dateItems.length - 2] + '.' + dateItems[dateItems.length - 1]
     dateItems.pop()
   }
-  const timeItems = dateItems[dateItems.length - 1].split(':')
-  if (ampmtoken !== undefined) {
-    timeItems.push(ampmtoken)
+
+  let parsedTime: Maybe<SimpleTime> = undefined
+  if (doesLookLikeATimeString(dateItems[dateItems.length - 1])) {
+    const timeItems = dateItems[dateItems.length - 1].split(TIME_SEPARATOR)
+    dateItems.pop()
+
+    parsedTime = defaultParseToTime(timeItems, timeFormat)
+    if (parsedTime === undefined) {
+      return undefined
+    }
   }
 
-  if (dateItems.length === 1) {
-    return defaultParseToTime(timeItems, timeFormat)
+  let parsedDate: Maybe<SimpleDate> = undefined
+  if (dateItems.length > 0) {
+    parsedDate = defaultParseToDate(dateItems, dateFormat)
+
+    if (parsedDate === undefined) {
+      return undefined
+    }
   }
-  if (timeItems.length === 1) {
-    return defaultParseToDate(dateItems, dateFormat)
-  }
-  const parsedDate = defaultParseToDate(dateItems.slice(0, dateItems.length - 1), dateFormat)
-  const parsedTime = defaultParseToTime(timeItems, timeFormat)
-  if (parsedDate === undefined) {
-    return undefined
-  } else if (parsedTime === undefined) {
-    return undefined
-  } else {
+
+  if (parsedDate && parsedTime) {
     return {...parsedDate, ...parsedTime}
   }
+
+  return parsedDate === undefined ? parsedTime : parsedDate
 }
 
 export const secondsExtendedRegexp = /^ss(\.(s+|0+))?$/
@@ -60,7 +66,7 @@ function defaultParseToTime(timeItems: string[], timeFormat: Maybe<string>): May
   } else if (timeFormat.endsWith('a/p')) {
     timeFormat = timeFormat.substring(0, timeFormat.length - 3).trim()
   }
-  const formatItems = timeFormat.split(':')
+  const formatItems = timeFormat.split(TIME_SEPARATOR)
   let ampm = undefined
   if (timeItems[timeItems.length - 1] === 'am' || timeItems[timeItems.length - 1] === 'a') {
     ampm = false
@@ -113,7 +119,7 @@ function defaultParseToDate(dateItems: string[], dateFormat: Maybe<string>): May
   if (dateFormat === undefined) {
     return undefined
   }
-  const formatItems = dateFormat.toLowerCase().split(/[ /.-]/g)
+  const formatItems = dateFormat.toLowerCase().split(DATE_SEPARATOR_REGEXP)
   if (dateItems.length !== formatItems.length) {
     return undefined
   }
@@ -161,4 +167,22 @@ function defaultParseToDate(dateItems: string[], dateFormat: Maybe<string>): May
   }
   const day = Number(dayString)
   return {year, month, day}
+}
+
+function doesLookLikeATimeString(s: string): boolean {
+  return s.includes(TIME_SEPARATOR)
+}
+
+function extractAmpmToken(s: string): Maybe<string> {
+  let ampmtoken: Maybe<string> = s.substring(s.length - 2)
+  if (ampmtoken === 'am' || ampmtoken === 'pm') {
+    return ampmtoken
+  }
+
+  ampmtoken = s.substring(s.length - 1)
+  if (ampmtoken === 'a' || ampmtoken === 'p') {
+    return ampmtoken
+  }
+
+  return undefined
 }