… CHANGELOG.md
… package.json
… src/DependencyGraph/Graph.ts
… src/DependencyGraph/ProcessableValue.ts
… src/DependencyGraph/TopSort.ts
… src/DependencyGraph/index.ts
… src/HyperFormula.ts
… src/Operations.ts
… src/errors.ts
… src/format/format.ts
… test/address-mapping.spec.ts
… test/column-range.spec.ts
… test/crud-random.spec.ts
… test/cruds/change-cell-content.spec.ts
… test/cruds/cut-paste.spec.ts
… test/cruds/move-cells.spec.ts
… test/cruds/removing-columns.spec.ts
… test/cruds/removing-rows.spec.ts
… test/graph-builder.spec.ts
… test/graph-garbage-collection.spec.ts
… test/graph-vertex.spec.ts
… test/graph.spec.ts
… test/named-expressions.spec.ts
… test/testUtils.ts
… test/volatile-functions.spec.ts