### Test case

A spreadsheet with 500k rows and 10 columns filled with string data. Total of 5M data cells, no formulas.

Script:
```ts
const hf = HyperFormula.buildFromArray([], {
  licenseKey: 'gpl-v3',
  maxRows: 500100,
  useStats: true,
  chooseAddressMappingPolicy: new AlwaysDense(),
})

const data = Array(500000).fill(0).map(() => Array(10).fill(0).map(() => 'A500000'))

var ty1 = (new Date()).getTime()
hf.setSheetContent(0, data)
var ty2 = (new Date()).getTime()
console.log(ty2 - ty1)
console.log(hf.getStats())
```

#### Ideas for improvement

- Function `getTopSortedWithSccSubgraphFrom`, which is the iterative implementation of the Tarjan algorithm that performs the topological sorting of the dependency graph and finds cycles (SCCs) in the graph. In this test case, the dependency graph is trivial; it contains only isolated nodes without any edges. It seems that it can be done more efficiently.
- Function `parseDateTimeFromConfigFormats`, which tries to parse all string data as date/time values. This test case contains no date/time values, so there might be some way of saving time by determining it quickly and avoiding running the heavy parsing operations.

### This PR focuses on optimizing topological sorting of the dependency graph

I made the Tarjan algorithm more efficient by changing the data structures it uses. Initially, the information about the graph nodes was stored in maps and sets with nodes as keys:

```ts
    const entranceTime: Map<T, number> = new Map()
    const low: Map<T, number> = new Map()
    const parent: Map<T, T> = new Map()
    const inSCC: Set<T> = new Set()
```
 
My approach was to use simple arrays indexed by integer ids and keep a single array of nodes as a mapping from id to node data.

```ts
  private nodes: T[] = []

  private entranceTime: number[] = []
  private low: number[] = []
  private parent: number[] = []
  private inSCC: boolean[] = []
```

### Results

Total time:
Before: **25923ms**
After: **11160ms**

Function `getTopSortedWithSccSubgraphFrom`:
Before: **51.6%**
After: **23.1%**

Profiler:
Before:
![image](https://github.com/handsontable/hyperformula/assets/2480086/1d26553f-a5f5-4dc4-9e5e-1f408e1f475c)

After:
![image](https://github.com/handsontable/hyperformula/assets/2480086/249b0a10-db02-4226-8a2f-6d908cd26ac2)

Profiler: Chrome Dev Tools

### Overall, I achieved the **56,02%** speed-up of HyperFormula for this use-case.

### How did you test your changes?
<!--- Describe in detail how you tested your changes. -->

- unit tests passed
- new unit tests for the Graph class

Comparison on our existing performance benchmarks on my PC:
```
                                     testName | before |  after
-----------------------------------------------------------------------
                                      Sheet A | 417.77 | 401.12
                                      Sheet B | 133.09 | 129.18
                                      Sheet T | 119.23 | 114.51
                                Column ranges | 292.47 | 408.56   <-- 40% slow-down
Sheet A:  change value, add/remove row/column |     23 |   27.2
 Sheet B: change value, add/remove row/column |    169 |  194.0
                   Column ranges - add column |  136.8 |  137.4
                Column ranges - without batch |  339.2 |  316.8
                        Column ranges - batch |    124 |   80.0
```

#### Column ranges benchmark

The 40% slow-down is observed only in Node environment. Running this benchmark in V8 yields similar results before and after applying the change.

### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in each box that applies. -->
- [ ] Breaking change (a fix or a feature because of which an existing functionality doesn't work as expected anymore)
- [x] New feature or improvement (a non-breaking change that adds functionality)
- [ ] Bug fix (a non-breaking change that fixes an issue)
- [ ] Additional language file, or a change to an existing language file (translations)
- [ ] Change to the documentation

### Related issues:
Fixes #876

### Checklist:
<!--- Go through the points below, and put an `x` in each box that applies. -->
<!--- If you're unsure about any of these, contact us. We're always glad to help! -->
- [ ] I have reviewed the guidelines about [Contributing to HyperFormula](https://hyperformula.handsontable.com/guide/contributing.html) and I confirm that my code follows the code style of this project.
- [ ] I have signed the [Contributor License Agreement](https://goo.gl/forms/yuutGuN0RjsikVpM2).
- [x] My change is compliant with the [OpenDocument](https://docs.oasis-open.org/office/OpenDocument/v1.3/os/part4-formula/OpenDocument-v1.3-os-part4-formula.html) standard.
- [x] My change is compatible with Microsoft Excel.
- [x] My change is compatible with Google Sheets.
- [x] I described my changes in the [CHANGELOG.md](https://github.com/handsontable/hyperformula/blob/master/CHANGELOG.md) file.
- [ ] My changes require a documentation update.
- [ ] My changes require a migration guide.