@sequba 2023-08-29
TODO:
- calculate the performance gain
- check other scenarios if we have no slow-downs (cruds?)
- try out other ideas from TODOs in the code