@sequba 2023-09-07
@budnix Could you also run the `Column ranges` benchmark yourself? I'm curios if you'd confirm my results. Here's the code:

```js
import { AlwaysDense, HyperFormula } from 'hyperformula';

function columnIndexToLabel(column) {
  let result = ''

  while (column >= 0) {
    result = String.fromCharCode((column % 26) + 97) + result
    column = Math.floor(column / 26) - 1
  }

  return result.toUpperCase()
}

function simpleCellAddressToString(address) {
  const column = columnIndexToLabel(address.col)
  return `${column}${address.row + 1}`
}

const cols = 50;
const data = [];
const firstRow = [1];

for (let i = 1; i < cols; ++i) {
  const adr = simpleCellAddressToString({sheet: 0, row: 0, col: i - 1});
  firstRow.push(`=${adr} + 1`);
}

data.push(firstRow);

for (let i = 1; i < cols; ++i) {
  const rowToPush = Array(i).fill(null);
  const startColumn = columnIndexToLabel(i - 1);

  for (let j = i - 1; j < cols - 1; ++j) {
    const endColumn = columnIndexToLabel(j);
    rowToPush.push(`=SUM(${startColumn}:${endColumn})`);
  }

  data.push(rowToPush);
}

const sheetId = 0;
const ty1 = (new Date()).getTime();

for (let i = 1; i < 200 ; ++i) {
  const hf = HyperFormula.buildFromArray([], {
    licenseKey: 'gpl-v3',
    maxRows: 500100,
    useStats: true,
    chooseAddressMappingPolicy: new AlwaysDense(),
  });

  hf.setSheetContent(sheetId, data);
}

const ty2 = (new Date()).getTime();
console.log(ty2 - ty1);
```