@budnix 2023-09-07
> @budnix Could you also run the `Column ranges` benchmark yourself? I'm curious if you'd confirm my results.

Using your code, the improvement isn't as spectacular as in PR's description. I tested on Node 20. Before PR ~25s after PR ~23s.