@AMBudnik 2023-09-19
Released in v2.6.0